+++ src/main/java/org/kdb/inside/brains/lang/formatting/blocks/AbstractQBlock.java
@@ -3,7 +3,6 @@ package org.kdb.inside.brains.lang.formatting.blocks;
 import com.intellij.formatting.*;
 import com.intellij.lang.ASTNode;
 import com.intellij.psi.TokenType;
-import com.intellij.psi.formatter.FormatterUtil;
 import com.intellij.psi.formatter.common.AbstractBlock;
 import com.intellij.psi.impl.source.tree.LeafPsiElement;
 import com.intellij.psi.tree.IElementType;
@@ -15,13 +14,16 @@ import org.kdb.inside.brains.psi.QTypes;
 import java.util.ArrayList;
 import java.util.List;
 
+import static org.kdb.inside.brains.lang.QNodeFactory.getFirstNotEmptyChild;
+import static org.kdb.inside.brains.lang.QNodeFactory.getNextNotEmptySibling;
+
 public abstract class AbstractQBlock extends AbstractBlock {
     protected static final Indent NONE_INDENT = Indent.getNoneIndent();
-
     protected final QFormatter formatter;
+    private final Indent indent;
+
     protected static final Indent SPACE_INDENT = Indent.getSpaceIndent(1);
     protected static final Indent NORMAL_INDENT = Indent.getNormalIndent();
-    private final Indent indent;
 
     public AbstractQBlock(@NotNull ASTNode node,
                           @NotNull QFormatter formatter) {
@@ -109,10 +111,6 @@ public abstract class AbstractQBlock extends AbstractBlock {
             return new TableBlock(node, formatter, wrap, alignment, indent);
         }
 
-        if (type == QTypes.FUNCTION_INVOKE_EXPR) {
-            return new FunctionInvokeBlock(node, formatter, wrap, alignment, indent);
-        }
-
         // One value wrapper type which should be expanded
         if (isWrapperType(type)) {
             return createBlock(getFirstNotEmptyChild(node), formatter, wrap, alignment, indent);
@@ -145,6 +143,7 @@ public abstract class AbstractQBlock extends AbstractBlock {
                 || type == QTypes.COLUMN_ASSIGNMENT_TYPE
                 || type == QTypes.SYMBOL
                 || type == QTypes.SYMBOLS
+                || type == QTypes.ITERATOR_TYPE
                 ;
     }
 
@@ -163,7 +162,6 @@ public abstract class AbstractQBlock extends AbstractBlock {
                 || type == QTypes.RETURN_EXPR
                 || type == QTypes.PREFIX_INVOKE_EXPR
                 || type == QTypes.PARENTHESES_INVOKE_EXPR
-                || type == QTypes.ITERATOR_TYPE
                 || type == QTypes.CONTEXT
                 || type == QTypes.CONTEXT_BODY
                 || type == QTypes.VAR_INDEXING
@@ -199,26 +197,6 @@ public abstract class AbstractQBlock extends AbstractBlock {
             child = getNextNotEmptySibling(child);
         }
         return result;
-    }
-
-    protected ASTNode getFirstNotEmptyChild(@NotNull ASTNode node) {
-        ASTNode res = node.getFirstChildNode();
-        while (res != null && isEmptyNode(res)) {
-            res = res.getTreeNext();
-        }
-        return res;
-    }
-
-    protected ASTNode getNextNotEmptySibling(@NotNull ASTNode node) {
-        ASTNode res = node.getTreeNext();
-        while (res != null && isEmptyNode(res)) {
-            res = res.getTreeNext();
-        }
-        return res;
-    }
-
-    protected boolean isEmptyNode(@NotNull ASTNode child) {
-        return FormatterUtil.containsWhiteSpacesOnly(child) || child.getTextLength() <= 0;
     }
 
     @FunctionalInterface